Fix microsec_clock() going negative at midnight

+++ src/gettod.c
@@ -311,7 +311,7 @@ int W32_CALL W32_NAMESPACE (gettimeofday) (struct timeval *tv, struct timezone *
 static __inline uint64 microsec_clock (void)
 {
   DWORD  hi;
-  long   lo;
+  WORD   lo;
   uint64 rc;
 
   do
